👌 IMPROVE: 重构 thinking-step-item 为模块化子组件
Made-with: Cursor
